+++ statisticssampler.h
@@ -6,27 +6,27 @@ class StatisticsSampler
 private:
     IO   *m_fileHandler;
     vec3  m_momentum;
-    float m_kineticEnergy;
-    float m_potentialEnergy;
-    float m_temperature;
-    float m_pressure;
-    float m_density;
+    double m_kineticEnergy;
+    double m_potentialEnergy;
+    double m_temperature;
+    double m_pressure;
+    double m_density;
 public:
     StatisticsSampler(IO *fileHandler);
     ~StatisticsSampler();
     void  sample(System *system);
-    float sampleKineticEnergy(System *system);
-    float samplePotentialEnergy(System *system);
-    float sampleTemperature(System *system);
-    float sampleDensity(System *system);
-    float samplePressure(System *system);
+    double sampleKineticEnergy(System *system);
+    double samplePotentialEnergy(System *system);
+    double sampleTemperature(System *system);
+    double sampleDensity(System *system);
+    double samplePressure(System *system);
     vec3 sampleMomentum(System *system);
 
     vec3 momentum() { return m_momentum; }
-    float totalEnergy();
-    float kineticEnergy() { return m_kineticEnergy; }
-    float potentialEnergy() { return m_potentialEnergy; }
-    float temperature() { return m_temperature; }
-    float pressure() { return m_pressure; }
-    float density() { return m_density; }
+    double totalEnergy();
+    double kineticEnergy() { return m_kineticEnergy; }
+    double potentialEnergy() { return m_potentialEnergy; }
+    double temperature() { return m_temperature; }
+    double pressure() { return m_pressure; }
+    double density() { return m_density; }
 };
